v1.22.0

+++ nerve/converter/converter-main/src/main/java/network/nerve/converter/tx/v1/HeterogeneousContractAssetRegCompleteProcessor.java
@@ -49,6 +49,7 @@ import network.nerve.converter.utils.ConverterSignValidUtil;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
 
+import static network.nerve.converter.constant.ConverterConstant.IN;
 import static network.nerve.converter.utils.ConverterUtil.addressToLowerCase;
 
 /**
@@ -91,6 +92,8 @@ public class HeterogeneousContractAssetRegCompleteProcessor implements Transacti
             Set<String> bindRemoveSet = new HashSet<>();
             Set<String> bindOverrideSet = new HashSet<>();
             Set<String> unregisterSet = new HashSet<>();
+            Set<String> pauseSet = new HashSet<>();
+            Set<String> resumeSet = new HashSet<>();
             for (Transaction tx : txs) {
                 byte[] coinData = tx.getCoinData();
                 if(coinData != null && coinData.length > 0){
@@ -100,7 +103,8 @@ public class HeterogeneousContractAssetRegCompleteProcessor implements Transacti
                 HeterogeneousContractAssetRegCompleteTxData txData = new HeterogeneousContractAssetRegCompleteTxData();
                 txData.parse(tx.getTxData(), 0);
                 String contractAddress = addressToLowerCase(txData.getContractAddress());
-                errorCode = ledgerAssetRegisterHelper.checkHeterogeneousContractAssetReg(chain, tx, contractAddress, txData.getDecimals(), txData.getSymbol(), txData.getChainId(), contractAssetRegSet, bindNewSet, bindRemoveSet, bindOverrideSet, unregisterSet, false);
+                errorCode = ledgerAssetRegisterHelper.checkHeterogeneousContractAssetReg(chain, tx, contractAddress, txData.getDecimals(), txData.getSymbol(), txData.getChainId(),
+                        contractAssetRegSet, bindNewSet, bindRemoveSet, bindOverrideSet, unregisterSet, pauseSet, resumeSet, false);
                 if (StringUtils.isNotBlank(errorCode)) {
                     failsList.add(tx);
                     continue;
@@ -137,7 +141,6 @@ public class HeterogeneousContractAssetRegCompleteProcessor implements Transacti
         Chain chain = null;
         try {
             chain = chainManager.getChain(chainId);
-            Map<Integer, List<HeterogeneousAssetInfo>> group = new HashMap<>();
             for (Transaction tx : txs) {
                 HeterogeneousContractAssetRegCompleteTxData txData = new HeterogeneousContractAssetRegCompleteTxData();
                 txData.parse(tx.getTxData(), 0);
@@ -149,7 +152,7 @@ public class HeterogeneousContractAssetRegCompleteProcessor implements Transacti
                 Integer hChainId = txData.getChainId();
                 IHeterogeneousChainDocking docking = heterogeneousDockingManager.getHeterogeneousDocking(hChainId);
                 // 异构合约资产注册 OR NERVE资产绑定异构合约资产: 新绑定 / 覆盖绑定
-                boolean isBindNew = false, isBindRemove = false, isBindOverride = false, isUnregister = false;
+                boolean isBindNew = false, isBindRemove = false, isBindOverride = false, isUnregister = false, isPause = false, isResume = false;
                 do {
                     byte[] remark = tx.getRemark();
                     if (remark == null) {
@@ -198,11 +201,45 @@ public class HeterogeneousContractAssetRegCompleteProcessor implements Transacti
                         isUnregister = true;
                         chain.getLogger().info("[commit] NERVE资产取消注册异构链合约资产[取消注册], NERVE asset: {}-{}, 异构链资产信息: chainId: {}, assetId: {}, symbol: {}, decimals: {}, address: {}",
                                 assetChainId, assetId, hChainId, info.getAssetId(), info.getSymbol(), info.getDecimals(), info.getContractAddress());
+                    } else if (BindHeterogeneousContractMode.PAUSE.toString().equals(mode)) {
+                        int hAssetId = 1;
+                        if (StringUtils.isNotBlank(info.getContractAddress())) {
+                            HeterogeneousAssetInfo assetInfo = docking.getAssetByContractAddress(info.getContractAddress());
+                            hAssetId = assetInfo.getAssetId();
+                        }
+                        info.setAssetId(hAssetId);
+                        String operationType = modeSplit[2];
+                        boolean isIn = IN.equalsIgnoreCase(operationType);
+                        if (isIn) {
+                            ledgerAssetRegisterHelper.pauseIn(hChainId, hAssetId);
+                        } else {
+                            ledgerAssetRegisterHelper.pauseOut(hChainId, hAssetId);
+                        }
+                        isPause = true;
+                        chain.getLogger().info("[commit] 异构链合约资产[充值暂停], NERVE asset: {}-{}, 异构链资产信息: chainId: {}, assetId: {}, symbol: {}, decimals: {}, address: {}",
+                                assetChainId, assetId, hChainId, info.getAssetId(), info.getSymbol(), info.getDecimals(), info.getContractAddress());
+                    } else if (BindHeterogeneousContractMode.RESUME.toString().equals(mode)) {
+                        int hAssetId = 1;
+                        if (StringUtils.isNotBlank(info.getContractAddress())) {
+                            HeterogeneousAssetInfo assetInfo = docking.getAssetByContractAddress(info.getContractAddress());
+                            hAssetId = assetInfo.getAssetId();
+                        }
+                        info.setAssetId(hAssetId);
+                        String operationType = modeSplit[2];
+                        boolean isIn = IN.equalsIgnoreCase(operationType);
+                        if (isIn) {
+                            ledgerAssetRegisterHelper.resumeIn(hChainId, hAssetId);
+                        } else {
+                            ledgerAssetRegisterHelper.resumeOut(hChainId, hAssetId);
+                        }
+                        isResume = true;
+                        chain.getLogger().info("[commit] 异构链合约资产[充值恢复], NERVE asset: {}-{}, 异构链资产信息: chainId: {}, assetId: {}, symbol: {}, decimals: {}, address: {}",
+                                assetChainId, assetId, hChainId, info.getAssetId(), info.getSymbol(), info.getDecimals(), info.getContractAddress());
                     }
 
                 } while (false);
                 // 异构合约资产注册
-                if (!isBindNew && !isBindRemove && !isBindOverride && !isUnregister) {
+                if (!isBindNew && !isBindRemove && !isBindOverride && !isUnregister && !isPause && !isResume) {
                     docking.saveHeterogeneousAssetInfos(List.of(info));
                     ledgerAssetRegisterHelper.crossChainAssetReg(chainId, hChainId, info.getAssetId(),
                             info.getSymbol(), info.getDecimals(), info.getSymbol(), info.getContractAddress());
@@ -240,7 +277,7 @@ public class HeterogeneousContractAssetRegCompleteProcessor implements Transacti
                 Integer hChainId = txData.getChainId();
                 IHeterogeneousChainDocking docking = heterogeneousDockingManager.getHeterogeneousDocking(hChainId);
                 // 异构合约资产注册 OR NERVE资产绑定异构合约资产: 新绑定 / 覆盖绑定
-                boolean isBindNew = false, isBindRemove = false, isBindOverride = false, isUnregister = false;
+                boolean isBindNew = false, isBindRemove = false, isBindOverride = false, isUnregister = false, isPause = false, isResume = false;
                 do {
                     byte[] remark = tx.getRemark();
                     if (remark == null) {
@@ -285,11 +322,42 @@ public class HeterogeneousContractAssetRegCompleteProcessor implements Transacti
                         ledgerAssetRegisterHelper.crossChainAssetReg(chainId, hChainId, info.getAssetId(),
                                 info.getSymbol(), info.getDecimals(), info.getSymbol(), info.getContractAddress());
                         isUnregister = true;
+                    } else if (BindHeterogeneousContractMode.PAUSE.toString().equals(mode)) {
+                        int hAssetId = 1;
+                        if (StringUtils.isNotBlank(info.getContractAddress())) {
+                            HeterogeneousAssetInfo assetInfo = docking.getAssetByContractAddress(info.getContractAddress());
+                            hAssetId = assetInfo.getAssetId();
+                        }
+                        info.setAssetId(hAssetId);
+                        String operationType = modeSplit[2];
+                        boolean isIn = IN.equalsIgnoreCase(operationType);
+                        if (isIn) {
+                            ledgerAssetRegisterHelper.resumeIn(hChainId, hAssetId);
+                        } else {
+                            ledgerAssetRegisterHelper.resumeOut(hChainId, hAssetId);
+                        }
+
+                        isPause = true;
+                    } else if (BindHeterogeneousContractMode.RESUME.toString().equals(mode)) {
+                        int hAssetId = 1;
+                        if (StringUtils.isNotBlank(info.getContractAddress())) {
+                            HeterogeneousAssetInfo assetInfo = docking.getAssetByContractAddress(info.getContractAddress());
+                            hAssetId = assetInfo.getAssetId();
+                        }
+                        info.setAssetId(hAssetId);
+                        String operationType = modeSplit[2];
+                        boolean isIn = IN.equalsIgnoreCase(operationType);
+                        if (isIn) {
+                            ledgerAssetRegisterHelper.pauseIn(hChainId, hAssetId);
+                        } else {
+                            ledgerAssetRegisterHelper.pauseOut(hChainId, hAssetId);
+                        }
+                        isResume = true;
                     }
 
                 } while (false);
                 // 异构合约资产注册
-                if (!isBindNew && !isBindRemove && !isBindOverride && !isUnregister) {
+                if (!isBindNew && !isBindRemove && !isBindOverride && !isUnregister && !isPause && !isResume) {
                     docking.rollbackHeterogeneousAssetInfos(List.of(info));
                     ledgerAssetRegisterHelper.deleteCrossChainAsset(hChainId, info.getAssetId());
                 }
